+++ src/Constants/HttpMethod.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * @author hollodotme
+ */
+
+namespace Fortuneglobe\IceHawk\Constants;
+
+/**
+ * Class Http
+ * @package Fortuneglobe\IceHawk\Constants
+ */
+abstract class HttpMethod
+{
+	const OPTIONS = 'OPTIONS';
+
+	const GET     = 'GET';
+
+	const HEAD    = 'HEAD';
+
+	const POST    = 'POST';
+
+	const PUT     = 'PUT';
+
+	const DELETE  = 'DELETE';
+
+	const TRACE   = 'TRACE';
+
+	const CONNECT = 'CONNECT';
+}