+++ src/Forms/FormData.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @author h.woltersdorf
+ * @author hollodotme
  */
 
 namespace Fortuneglobe\IceHawk\Forms;