+++ tests/Unit/Fixtures/TestUriRewriterWithValidMap.php
@@ -5,23 +5,22 @@
 
 namespace Fortuneglobe\IceHawk\Tests\Unit\Fixtures;
 
-use Fortuneglobe\IceHawk\Constants\Http;
+use Fortuneglobe\IceHawk\Constants\HttpCode;
+use Fortuneglobe\IceHawk\Defaults\UriRewriter;
 use Fortuneglobe\IceHawk\Interfaces\ServesRequestInfo;
-use Fortuneglobe\IceHawk\UriRewriter;
 
 /**
  * Class TestUriRewriterWithValidMap
- *
  * @package Fortuneglobe\IceHawk\Tests\Unit\Fixtures
  */
 class TestUriRewriterWithValidMap extends UriRewriter
 {
 
 	private static $simpleMap = [
-		"#/non/regex/rewrite#"      => [ '/non_regex_rewrite', Http::MOVED_PERMANENTLY ],
+		"#/non/regex/rewrite#"      => [ '/non_regex_rewrite', HttpCode::MOVED_PERMANENTLY ],
 		"#/non/regex/no/code#"      => [ '/non_regex_no_code' ],
-		"#^/regex/rewrite/?#"       => [ '/regex_rewrite', Http::MOVED_TEMPORARILY ],
-		"#^/regex/param/([^/]+)/?#" => [ '/regex_param_$1', Http::MOVED_TEMPORARILY ],
+		"#^/regex/rewrite/?#"       => [ '/regex_rewrite', HttpCode::FOUND ],
+		"#^/regex/param/([^/]+)/?#" => [ '/regex_param_$1', HttpCode::FOUND ],
 	];
 
 	public function rewrite( ServesRequestInfo $requestInfo )